# docs/api/signalville-openapi.yaml
openapi: 3.1.0
info:
  title: API SignalVille
  version: 1.1.0
  description: >
    API REST de la plateforme SignalVille, destinée à la gestion des incidents urbains,
    des utilisateurs, des interventions, des notifications et des statistiques.
servers:
  - url: http://localhost:8080/api/v1
    description: Environnement local
...
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      tags: [Authentification]
      summary: Inscrire un citoyen
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Compte citoyen créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Adresse e-mail déjà utilisée
  /auth/login:
    post:
      tags: [Authentification]
      summary: Authentifier un utilisateur
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentification réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Authentification]
      summary: Rafraîchir le token d'accès
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Nouveau token d'accès émis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [Authentification]
      summary: Déconnexion (révocation du refresh token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '204':
          description: Déconnexion effectuée, refresh token révoqué
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /categories:
    get:
      tags: [Catégories]
      summary: Lister les catégories actives
      security: []
      responses:
        '200':
          description: Liste des catégories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'
    post:
      tags: [Catégories]
      summary: Créer une catégorie
      description: Réservé aux administrateurs.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '201':
          description: Catégorie créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
  /categories/all:
    get:
      tags: [Catégories]
      summary: Lister toutes les catégories, actives et inactives
      description: >
        Réservé aux administrateurs. Contrairement à GET /categories (public,
        actives uniquement), cette route expose l'intégralité du catalogue —
        nécessaire pour retrouver et réactiver une catégorie désactivée par
        erreur.
      responses:
        '200':
          description: Liste complète des catégories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /categories/{categoryId}:
    get:
      tags: [Catégories]
      summary: Consulter une catégorie
      description: >
        Réservé aux administrateurs. Retourne la catégorie quel que soit son
        statut (active ou inactive), contrairement à GET /categories qui ne
        liste que les actives.
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      responses:
        '200':
          description: Catégorie trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Catégories]
      summary: Modifier une catégorie
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '200':
          description: Catégorie modifiée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
  /categories/{categoryId}/status:
    patch:
      tags: [Catégories]
      summary: Activer ou désactiver une catégorie
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [active]
              properties:
                active:
                  type: boolean
      responses:
        '204':
          description: Statut modifié
...
  /reports/{reportId}/notes:
    get:
      tags: [Interventions]
      summary: Lister les notes visibles
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: Notes du signalement
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NoteResponse'
    post:
      tags: [Interventions]
      summary: Ajouter une note
      description: Réservé aux rôles AGENT et SUPERVISEUR. Le citoyen n'a pas accès à cet endpoint.
      parameters:
        - $ref: '#/components/parameters/ReportId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNoteRequest'
      responses:
        '201':
          description: Note ajoutée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteResponse'
  /interventions/{interventionId}/proofs:
    get:
      tags: [Interventions]
      summary: Lister les preuves de resolution d'une intervention
      parameters:
        - $ref: '#/components/parameters/InterventionId'
      responses:
        '200':
          description: Liste des preuves de l'intervention
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PhotoResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /interventions/{interventionId}/proofs/{proofId}:
    get:
      tags: [Interventions]
      summary: Telecharger le binaire d'une preuve de resolution
      parameters:
        - $ref: '#/components/parameters/InterventionId'
        - name: proofId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Fichier binaire de la preuve (image)
          content:
            image/*:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    CategoryId:
      name: categoryId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ReportId:
      name: reportId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    InterventionId:
      name: interventionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentification requise ou invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Accès interdit
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    RoleName:
      type: string
      enum: [CITOYEN, AGENT, SUPERVISEUR, ADMINISTRATEUR]
    AccountStatus:
      type: string
      enum: [ACTIF, SUSPENDU, DESACTIVE]
...
    Priority:
      type: string
      enum: [BASSE, MOYENNE, HAUTE, CRITIQUE]
    NoteType:
      type: string
      enum: [INTERNE, PUBLIC]
    RegisterRequest:
      type: object
      required: [firstName, lastName, email, phone, password, confirmPassword, termsAccepted]
      properties:
        firstName:
          type: string
          minLength: 2
        lastName:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          minLength: 8
        confirmPassword:
          type: string
        termsAccepted:
          type: boolean
          const: true
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
        user:
          $ref: '#/components/schemas/UserResponse'
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        role:
          $ref: '#/components/schemas/RoleName'
        status:
          $ref: '#/components/schemas/AccountStatus'
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: [string, 'null']
          format: date-time
...
    CategoryRequest:
      type: object
      required: [name, defaultPriority, targetDelayHours]
      properties:
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        defaultPriority:
          $ref: '#/components/schemas/Priority'
        targetDelayHours:
          type: integer
          minimum: 1
        active:
          type: boolean
          default: true
    CategoryResponse:
      allOf:
        - $ref: '#/components/schemas/CategoryRequest'
        - type: object
          properties:
            id:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
...
    PhotoResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        description:
          type: string
        order:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    CreateNoteRequest:
      type: object
      required: [content, type]
      properties:
        content:
          type: string
        type:
          $ref: '#/components/schemas/NoteType'
    NoteResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        type:
          $ref: '#/components/schemas/NoteType'
        author:
          $ref: '#/components/schemas/UserResponse'
        createdAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        validationErrors:
          type: object
          additionalProperties:
            type: string